good_chewingDj/requirements.txt
line-bot-sdk==2.3.0
flask==2.1.2
requests==2.28.0
speechrecognition==3.8.1
pydub==0.25.1